+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/graph/BaseBindingStack.kt
@@ -9,6 +9,7 @@ internal interface BaseBindingStack<
   Type : Any,
   TypeKey : BaseTypeKey<Type, *, *>,
   Entry : BaseBindingStack.BaseEntry<Type, TypeKey, *>,
+  Impl : BaseBindingStack<ClassType, Type, TypeKey, Entry, Impl>,
 > {
   val graph: ClassType
   val entries: List<Entry>
@@ -17,6 +18,8 @@ internal interface BaseBindingStack<
   fun push(entry: Entry)
 
   fun pop()
+
+  fun copy(): Impl
 
   fun entryFor(key: TypeKey): Entry?