+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/BindingGraphGenerator.kt
@@ -34,6 +34,7 @@ internal class BindingGraphGenerator(
     val graph =
       IrBindingGraph(
         this,
+        node,
         newBindingStack = {
           IrBindingStack(node.sourceGraph, loggerFor(MetroLogger.Type.BindingGraphConstruction))
         },