+++ compiler/src/test/kotlin/dev/zacsweers/metro/compiler/transformers/GraphExtensionTest.kt
@@ -560,7 +560,7 @@ class GraphExtensionTest : MetroCompilerTest() {
     compile(
       source(
         """
-          @DependencyGraph
+          @DependencyGraph(AppScope::class)
           abstract class ExampleGraph {
             @Provides
             @SingleIn(AppScope::class)